+++ packages/create-app/src/blank-webapp/static/src/pages/second/second-page.ts
@@ -1,5 +1,5 @@
 import { html, LitElement } from 'lit';
-import { PageController } from '@fork-cells/page-controller';
+import { PageController } from '@andresbitrian/page-controller';
 import { customElement } from 'lit/decorators.js';
 
 @customElement('second-page')